backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
ultralytics>=8.2.0,<9.0.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pyyaml==6.0.1
reportlab==4.1.0
Pillow==10.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0
pytest-asyncio==0.23.6
aiofiles==23.2.1
